add react hook eslint plugin
