+++ src/internal/useOptionalSekai.ts
@@ -2,13 +2,12 @@ import { useContext } from 'react'
 
 import { YourSekaiContext } from '@/components/provider/YourSekaiProvider'
 
-import { useCreateSekai } from '@/hooks/useCreateSekai'
 import { LIGHT_MODE } from '@/hooks/useThemeMode'
 
 import { COLORS_SEKAI_KEYS, colorsSekai } from '@/styles/sekai-colors'
 
 import type { PaletteMode } from '@/hooks/useThemeMode'
-import type { ColorsSekaiKey } from '@/styles/sekai-colors'
+import type { ColorsSekaiCode, ColorsSekaiKey } from '@/styles/sekai-colors'
 import type { SekaiTheme } from '@/utils/createSekai'
 
 const defaultTheme: SekaiTheme = {
@@ -28,12 +27,13 @@ type OptionalSekaiProps = {
 
 export const useOptionalSekai = (option: OptionalSekaiProps) => {
   const context = useContext(YourSekaiContext)
-  const sekaiColor = context
-    ? colorsSekai[option.sekai || useCreateSekai().sekaiTheme.palette.sekai]
+
+  const sekaiColor: ColorsSekaiCode = context
+    ? colorsSekai[option.sekai || context.sekaiTheme.palette.sekai]
     : colorsSekai[defaultTheme.palette.sekai]
 
   const modeTheme = context
-    ? option.mode || useCreateSekai().sekaiTheme.palette.mode
+    ? option.mode || context.sekaiTheme.palette.mode
     : defaultTheme.palette.mode
 
   const isLight = LIGHT_MODE === modeTheme
